+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - [BREAKING] Refactor transaction kernel advice inputs (#1425).
 - Remove miden-proving-service binary crate and miden-proving-service-client crate (#1427).
 - Remove doc update checks on CI (#1435).
-- Introduce `ScriptMastForestStore` and refactor MAST forest provisioning in the `TransactionExecutor` (#1438).
+- [BREAKING] Introduce `ScriptMastForestStore` and refactor MAST forest provisioning in the `TransactionExecutor` (#1438).
 
 ## 0.9.1 (2025-05-30)
