+++ crates/miden-tx/src/host/script_mast_forest_store.rs
